switching over to snapshot time in milliseconds.
git-svn-id: svn://activequant.com/svn/repositories/aq2o/trunk@184 01b8edfa-0b2f-4ad0-b428-228ee2dd3b63

+++ src/main/java/com/activequant/domainmodel/PersistentEntity.java
@@ -16,7 +16,7 @@ public abstract class PersistentEntity {
     private String className;
     private long creationTime = System.currentTimeMillis();
     private long deletionTime = 0L;
-    private double snapshotTime = 0.0;
+    private long snapshotTime = 0L;
     private static Logger log = Logger.getLogger(PersistentEntity.class);
     private Map<String, Object> underlyingMap = new HashMap<String, Object>();
 
@@ -50,12 +50,11 @@ public abstract class PersistentEntity {
     public abstract String getId();
 
     /**
-     * Date8Time6.
-     * 
-     * @return when this value was stored, in date8time6 format.
+     *  
+     * @return when this value was stored, in ordinary milliseconds. 
      */
     @Property
-    public double getSnapshotTime() {
+    public long getSnapshotTime() {
         return snapshotTime;
     }
 
@@ -212,7 +211,7 @@ public abstract class PersistentEntity {
         this.deletionTime = deletionTime;
     }
 
-    public void setSnapshotTime(double snapshotTime) {
+    public void setSnapshotTime(long snapshotTime) {
         this.snapshotTime = snapshotTime;
     }
